+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to the "SnapTeX" extension will be documented in this file.
 
+## [0.5.3] - 2026-01-09
+- **Fix bug**: webview async image canvas func load
+
 ## [0.5.2] - 2026-01-09
 - **Fix**: image 401 error